Reduced the thread count, increased the buffer size

+++ src/main/java/fogetti/phish/storm/integration/PhishTopologyRemoteRunner.java
@@ -14,11 +14,11 @@ public class PhishTopologyRemoteRunner {
 		config.setNumWorkers(20);
 		config.setMessageTimeoutSecs(60);
 		config.put(Config.TOPOLOGY_EXECUTOR_RECEIVE_BUFFER_SIZE,
-	               new Integer(4096));
+	               new Integer(262144));
 	    config.put(Config.TOPOLOGY_EXECUTOR_SEND_BUFFER_SIZE,
-	               new Integer(4096));
+	               new Integer(262144));
 	    config.put(Config.TOPOLOGY_TRANSFER_BUFFER_SIZE,
-	               new Integer(4096));
+	               new Integer(262144));
 	    config.registerMetricsConsumer(LoggingMetricsConsumer.class, 1);
 		
 		StormSubmitter.submitTopology("phish-storm-topology", config, topology);
